+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Changed
+
+- Improve portfolio pie chart responsivness
+
 ## [0.2.0] − 2019-03-05
 
 ### Added